Refactor PollDisplay + children
So far, PollDisplay, PollHeader, and RespondentSelector
have been refactored into functional components, stored
in separate files.
