+++ app/javascript/components/PollDisplay/QuestionsContainer.jsx
@@ -1,0 +1,22 @@
+import React from "react";
+import QuestionDisplay from "./QuestionDisplay"
+
+export default class QuestionsContainer extends React.Component {
+  // props: questions, response_options
+  render () {
+    const questions = this.props.questions
+    const response_options = this.props.response_options
+    const questionDisplayListItems = questions.map((question, index) =>
+      <QuestionDisplay
+        key = {index}
+        question = {question}
+        response_options = {response_options[question.id]}
+      />
+    )
+    return (
+      <ul className = "questions-container">
+        {questionDisplayListItems}
+      </ul>
+    )
+  }
+}
